go: Apply package exclusion equally
The exclusion logic was moved out of getGoDirs but only buildPhase was updated
causing checkPhase to possibly fail. This happened in golint as it has go
files in testdata that are meant as testdata files and not go packages to
test which caused the checkPhase to fail.

This is a port of https://github.com/NixOS/nixpkgs/commit/ddcc7c67d280eb0ea76a9079c49c9f70c6cc2e4f