Enter vendor directories recursively when symlinking fails
There could be several levels of already-existing directories.

An example of this is version 0.6.0 of guac
(https://github.com/guacsec/guac), where there is
`github.com/aws/aws-sdk-go-v2/aws` and several
`github.com/aws/aws-sdk-go-v2/aws/protocol/*`, but no
`github.com/aws/aws-sdk-go-v2/aws/protocol`.

There is no risk of infinite recursion here because it goes down
the directory hierarchy one level each recursive call, and will
terminate when no directory is found here anymore. The exception
to this is when both the existing tree and the new vendored
dependency have the same symlink loop, but that seems unlikely.

It would be good to have a unit test for this behaviour, but it's
been a while since I did go seriously, so I'd like to leave it at
this.