fix bug around go version 1.22
Any evaluation of the go_1_22 package will throw an exception.
This was causing the `selectGo` function to fail with a decryptic
message saying that go 1.22 was not supported, even when there was
no reference at all of it anywhere.

This changes will ensure that when we loop through the available
go packages, we test them against any `throw` alias - which are
used for deprecation.